pyproject.toml
[project]
name = "whatsapp-crew"
version = "0.1.0"
description = "A CrewAI project for summarizing WhatsApp group activity"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "crewai",
    "crewai-tools",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "pytz>=2024.1",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.1",
    "regex>=2023.12.25"
]

[project.scripts]
whatsapp_crew = "whatsapp_crew.main:run"
run_crew = "whatsapp_crew.main:run"
train = "whatsapp_crew.main:train"
replay = "whatsapp_crew.main:replay"
test = "whatsapp_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.4",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.1",
    "isort>=5.13.2",
    "mypy>=1.8.0"
]
